+++ doc/changelog.rst
@@ -9,3 +9,9 @@ Version 0.2 (*under development*)
 
 - removed redundant method :meth:`nd.classify.cluster()`, as same
   functionality can be achieved using :class:`nd.classify.Classifier`
+
+:mod:`nd.tiling`
+..................
+
+- added :meth:`nd.tiling.debuffer()` to automatically remove buffer from
+  tiled datasets